pdf-parser-service/requirements.txt
Django
djangorestframework
PyPDF2
pdfminer.six
requests
numpy
scikit-learn
pandas